+++ frontend/src/components/Task/displaytext.tsx
@@ -1,3 +1,4 @@
+import { Box } from '@chakra-ui/react';
 import { observer } from 'mobx-react-lite';
 
 import { TaskType } from '@/lib/models/Task';
@@ -9,9 +10,19 @@ export const DisplayText: React.VFC<{
 }> = observer(({ task, width, height }) => {
   return (
     <foreignObject x="0" y="0" width={width} height={height} fill="none">
-      <div style={{ textAlign: 'center', padding: '4px' }}>
+      <Box
+        style={{
+          width: '100%',
+          height: '100%',
+          textAlign: 'center',
+          padding: '4px',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center',
+        }}
+      >
         <span>{task.name}</span>
-      </div>
+      </Box>
     </foreignObject>
   );
 });
