+++ frontend/src/components/Task/process.tsx
@@ -28,7 +28,12 @@ export const Process: React.VFC<{ task: TaskType }> = observer(({ task }) => {
       />
       <DisplayText task={task} width={size.width} height={size.height} />
       {task.label && (
-        <Label task={task} width={size.width} height={size.height} />
+        <Label
+          task={task}
+          width={size.width}
+          height={size.height}
+          color="#7936CD"
+        />
       )}
     </g>
   );
